+++ src/index.ts
@@ -5,5 +5,3 @@ export * from "./matchers";
 export * from "./comparison";
 
 export * from "./equality";
-
-export * from "./staticEquality";