+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Version 4
+
+- `ClassWithStaticEquals` is no more exported
+
+- `StaticEquality` exists no more; in particular, its `addTesterFor` now belongs to the `Equality` namespaces - and client classes must turn their `static equals(left, right)` into a non-static `equals(other)` method.
+
 ## Version 3
 
 - `.toBeSameSequence()` renamed to `.toEqualSequence()`